测试asset的startup 通过

+++ scripts/asset_publish/src/submit/asset.py
@@ -2,6 +2,9 @@
 import json
 import os
 
+from PySide2.QtWidgets import QMessageBox
+
+from m_maya_py2.src.file import MayaFileOperatorPy2, SaveType
 from scripts.asset_publish.src.submit import SubmitDataABC
 from scripts.asset_publish.src.submit.public import PublicData
 from scripts.asset_publish.src.view.check import CheckView
@@ -9,7 +12,8 @@ from scripts.cache_path.task import CacheSubmitTaskStrategy
 
 
 class AssetData(object):
-    def __init__(self, asset_type, asset_name, task_name, artist, pipeline, task_id):
+    def __init__(self, project_db, asset_type, asset_name, task_name, artist, pipeline, task_id):
+        self.project_db = project_db
         self.asset_type = asset_type
         self.asset_name = asset_name
         self.task_name = task_name
@@ -28,6 +32,7 @@ class AssetData(object):
 
     def to_dict(self):
         return {
+            'project_db': self.project_db,
             'asset_type': self.asset_type,
             'asset_name': self.asset_name,
             'task_name': self.task_name,
@@ -38,8 +43,8 @@ class AssetData(object):
 
 
 class AssetSubmit(PublicData, SubmitDataABC):
-    def __init__(self, project_db, asset_data):
-        super(AssetSubmit, self).__init__(project_db, "asset",
+    def __init__(self, asset_data):
+        super(AssetSubmit, self).__init__(asset_data.project_db, "asset",
                                           CacheSubmitTaskStrategy("ZMPublish", asset_data.asset_name))
         self.asset_data = asset_data
         if not os.path.exists(self.path):
@@ -50,14 +55,22 @@ class AssetSubmit(PublicData, SubmitDataABC):
 
     def submit(self, parent=None):
         # 生成json文件
+        maya_file = MayaFileOperatorPy2()
+        if not maya_file.path:
+            QMessageBox.critical(None, "Error", u"请保存工程后再进行操作")
+            return
         data = self.asset_data.to_dict()
-        data["project_db"] = self.project_db
+        data["project_name"] = self.project_db.replace("proj_", "")
         data["module"] = "asset"
         data["window_name"] = "[{}]{}".format(data.get("task_name"), data.get("asset_name"))
+        # 写入缓存
         with open(self.full_path, "w") as f:
             json.dump(data, f)
+        # 写入文件信息
+        maya_file.add_file_info(data)
         self.check_view = CheckView(parent)
         self.check_view.run(data)
+        maya_file.save(SaveType.ma)
 
 # if __name__ == '__main__':
 #     a = AssetDataStrategy("proj_csx2", AssetData("prop", "my_test", "body", "wangshuo", "123-321"))
